Removing dependency on Oculus
Added changed quite a few things to allow support for running skyline
without Oculus, and to stop writing to the mini namespace in redis.

The horizon worker now checks to see if the OCULUS_HOST is empty, and
if it is, then it doesn't write to the mini namespace.

The webapp no longer uses the mini namespace to populate data for the
1 hour graph, and uses a slice of the data from the full namespace
returned from the api to get that data. As such, we no longer need
the mini namespace in the .js file or to send the mini namespace from
the python webapp.py. It also doesn't append a link to Oculus if the
OCULUS_HOST setting is an empty string. The settings.py.example has
been changed to reflect this.