Removing dependency on Oculus
Got overzealous when cutting out stuff in skyline.js and cut out a
metric when I didn't mean to. Also moved the logic to skip the mini
namespace into the horizon-agent.py since roomba and worker can both use
it, then passed it into the constructors for both of them to use.